Embedded: Load embedded model in represetation
.. to allow preview in edit mode.
Embedded Widget runtime only starts/stops the embedded model, it's
already loaded and represented by the embedded widget representation.

#105
