+++ org.csstudio.display.builder.runtime/src/org/csstudio/display/builder/runtime/ActionUtil.java
@@ -16,6 +16,7 @@ import org.csstudio.display.builder.model.DisplayModel;
 import org.csstudio.display.builder.model.Widget;
 import org.csstudio.display.builder.model.macros.MacroHandler;
 import org.csstudio.display.builder.model.macros.Macros;
+import org.csstudio.display.builder.model.persist.ModelLoader;
 import org.csstudio.display.builder.model.properties.ActionInfo;
 import org.csstudio.display.builder.model.properties.ExecuteScriptActionInfo;
 import org.csstudio.display.builder.model.properties.OpenDisplayActionInfo;
@@ -77,7 +78,7 @@ public class ActionUtil
             final String parent_file = widget_model.getUserData(DisplayModel.USER_DATA_INPUT_FILE);
 
             // Load new model. If that fails, no reason to continue.
-            final DisplayModel new_model = RuntimeUtil.loadModel(parent_file, expanded_path);
+            final DisplayModel new_model = ModelLoader.loadModel(parent_file, expanded_path);
 
             // Model is standalone; source_widget (Action button, ..) is _not_ the parent,
             // but it does add macros to those already defined in the display file.
